Gridding updates and more detailed flux comparison
The aim point gridding was updated to enable grid
definitions the have the aim points as a subset
of the measurement points.

Numerical experiments (and supporting utilities)
were also added to explore the impact on the
measurement point/ aimpoint grids on the computed
total power and peak fluxes in SolarPILOT vs.
HALOS's shifting method.